remove cruft, move config to app, cleanup build task
